# src/main/java/org/tkit/parameters/rs/internal/dtos/ApplicationParameterPageResultDTO.java
package org.tkit.parameters.rs.internal.dtos;

import io.quarkus.runtime.annotations.RegisterForReflection;

@RegisterForReflection
public class ApplicationParameterPageResultDTO extends PageResultDTO<ApplicationParameterDTO> {

}